==> 01_gerar_dados.ipynb <==
import gol_apis as apis
from api_tools import setup as stp
from gol_logger import Logger

import pandas as pd
import numpy as np


from modules.generate_data import get_data_neolude
from modules.generate_data import get_rosters
from modules.generate_data import get_crew_info
from modules.generate_data import get_checks

import sys
from config.variables import user

Relação de Treinamentos mapeados do **NEOLUDE**:
1.  14140 : ARP, TEC
2.  6167  : CFI, TEC
3.  16766 : EQP, TEC
4.  18180 : EMG, TEC
5.  25250 : EMG_presencial, TEC
6.  14241 : SGSO TRIPULANTES, TEC e COM
7.  17877 : OPT, TEC
8.  11211 : RBAC, TEC e COM
9.  26361 : aproximacoes_estabilizadas_e_nao_estabilizadas, TEC
10. 19897 : experiencia_recente, TEC
11. 6069  : CRM, TEC e COM
12. 6162  : TEM, TEC
13. 6173  : SEC, TEC
14. 6066  : SEC_inicial, TEC
15. 14951 : ARP_tripulacao_comercial, COM
16. 14547 : PERIODICO II - REVALIDAÇÃO - RVB - MÓDULO EAD, COM
17. 25352 : PERIODICO II - REVALIDACAO - RVB - MÓDULO PRÁTICO, COM
18. 14646 : PERIODICO I - RECICLAGEM - RCL - MÓDULO EAD, COM
19. 26463 : PERIODICO I - RECICLAGEM - RCL - MÓDULO PRÁTICO, COM
20. 14951 : DGR - CHAVE 11 - PERIODICO - EAD, COM
21. 27775 : DGR - CHAVE 11 - INICIAL (Artigos Perigosos - chave 11 - Inicial), COM
22. 20301 : CPEX - Curso Periódico de Examinador - EAD, COM
23. 20907 : CPIR - Curso Periodico de Instrutor de Voo - EAD, COM
24. 31310 : Treinamento Anticorrupção V. 2021, TEC e COM
25. 2425  : O Terrorismo na Aviação Civil, TEC e COM
26. 6066  : AVSEC PARA TRIPULANTES - FORMAÇÃO RBAC 110, COM
27. 6173  : AVSEC PARA TRIPULANTES - ATUALIZAÇÃO RBAC 110 COM

# Gerar dados de roster do mes
get_rosters(user, log_info=False, period=dt(2021,8,1), all_period=False)

# Gerar dados de Checks
get_checks(user, log_info=False, active=True)

# Gerar dados de Crew Info
get_crew_info(user, log_info=False, date=dt.now())

# Gerar dados de ARP
get_data_neolude(user, log_info=False, training_id=14140)

# Gerar dados de CFI
get_data_neolude(user, log_info=False, training_id=6167)

# Gerar dados de EQP
get_data_neolude(user, log_info=False, training_id=16766)

# Gerar dados de EMG 
get_data_neolude(user, log_info=False, training_id=18180)

# Gerar dados de EMG presencial
get_data_neolude(user, log_info=False, training_id=25250)

# Gerar dados de SGSO
get_data_neolude(user, log_info=False, training_id=14241)

# Gerar dados de CRM
get_data_neolude(user, log_info=False, training_id=6069)

# Gerar dados de C-TEM
get_data_neolude(user, log_info=False, training_id=6162)

# Gerar dados de SEC
get_data_neolude(user, log_info=False, training_id=6173)

# Gerar dados de SEC-INICIAL
get_data_neolude(user, log_info=False, training_id=6066)

# Gerar dados de ARP tripulação comercial
get_data_neolude(user, log_info=False, training_id=14951)

# Gerar dados de Fadiga tripulação comercial
get_data_neolude(user, log_info=False, training_id=11211)

# Gerar dados do treinamento de ANTI CORRUPÇÃO
get_data_neolude(user, log_info=False, training_id=31310)

# Gerar dados do treinamento de ANTI CORRUPÇÃO
get_data_neolude(user, log_info=False, training_id=14547)

# Gerar dados do treinamento de ANTI CORRUPÇÃO
get_data_neolude(user, log_info=False, training_id=25352)

# Gerar dados do treinamento de ANTI CORRUPÇÃO
get_data_neolude(user, log_info=False, training_id=14646)

# Gerar dados do treinamento de ANTI CORRUPÇÃO
get_data_neolude(user, log_info=False, training_id=26463)

# Gerar dados do treinamento de ANTI CORRUPÇÃO
get_data_neolude(user, log_info=False, training_id=14951)

# Gerar dados do treinamento de ANTI CORRUPÇÃO
get_data_neolude(user, log_info=False, training_id=27775)

# Gerar dados do treinamento de ANTI CORRUPÇÃO
get_data_neolude(user, log_info=False, training_id=20301)

# Gerar dados do treinamento de ANTI CORRUPÇÃO
get_data_neolude(user, log_info=False, training_id=20907)

# Gerar dados do treinamento de ANTI CORRUPÇÃO
get_data_neolude(user, log_info=False, training_id=14241)

# Gerar dados do treinamento de ANTI CORRUPÇÃO
get_data_neolude(user, log_info=False, training_id=2425)

# Gerar dados do treinamento de ANTI CORRUPÇÃO
get_data_neolude(user, log_info=False, training_id=6173)

# Gerar dados do treinamento de CFI Inicial
get_data_neolude(user, log_info=False, training_id=6167)
