==> 06_EMG.ipynb <==
from gol_apis import RosterActivitiesService
from gol_apis import RosterTagService
from gol_apis import CheckScheduleService
from gol_apis import CrewMemberInfoService
import api_tools
from api_tools.setup import setup_crew_info_dataframe
from api_tools.setup import setup_checks_dataframe
from api_tools.setup import setup_tag_dataframe
from api_tools.setup import setup_activities_dataframe

from gol_logger import Logger
import pandas as pd
import multiprocessing as mp
import requests, time
from datetime import datetime as dt
from datetime import timedelta as td
from utils import funcs
from utils.progress import step
from utils.progress import BAR
from itertools import chain

from modules.generate_data import get_data_neolude
from modules.generate_data import get_rosters
from modules.generate_data import get_crew_info
from modules.generate_data import get_checks

import sys
from config.variables import user

training_code = 'EMG'

# Dados Crew Info
crew = pd.read_csv('data/crew/crew_info.csv', sep=';', encoding='utf-8-sig')
crew.cif = crew.cif.astype(int).astype(str).str.zfill(8)
crew = crew[crew['funcao'].isin(['CMT', 'COP'])]
crew = crew[['cif', 'guerra', 'funcao', 'cargo', 'base']]
crew

# Header para filtrar dados do Neolude
header_neolude = ['RegistrationNumber',
                  'UserName',
                  'BusinessUnitName',
                  'Post',
                  'CPF',
                  'Email',
                  'CourseName',
                  'CourseUserStatus',
                  'GradeFormatted',
                  'LastAccessDate',
                  'CreatedDate',
                  'CompletionDate']

# Dados Neolude
training = pd.read_csv(f'data/neolude/{training_code}.csv', sep=';', encoding='utf-8-sig')
training = training[header_neolude]
training['tipo'] = pd.to_numeric(training['RegistrationNumber'], errors='coerce').notnull()
training = training[training['tipo']]
training['RegistrationNumber'] = training['RegistrationNumber'].fillna(0)
training['RegistrationNumber'] = training['RegistrationNumber'].astype(int).astype(str).str.zfill(8)
training['LastAccessDate'] = pd.to_datetime(training['LastAccessDate'])
training['CreatedDate'] = pd.to_datetime(training['CreatedDate'])
training['CompletionDate'] = pd.to_datetime(training['CompletionDate'])
training.columns = ['cif', 'name', 'base', 'funcao', 'cpf', 'email', 'course_name','status', 'nota', 'ultimo_acesso', 'criacao_acesso', 'data_conclusao', 'tipo']
training = training[['cif', 'course_name','status', 'nota', 'data_conclusao']]
training = training.sort_values(by=['cif','data_conclusao'])
training = training.drop_duplicates(subset = ['cif'], keep = 'last')
training

# Dados Neolude
training_p = pd.read_csv(f'data/neolude/{training_code}_presencial.csv', sep=';', encoding='utf-8-sig')
training_p = training_p[header_neolude]
training_p['tipo'] = pd.to_numeric(training_p['RegistrationNumber'], errors='coerce').notnull()
training_p = training_p[training_p['tipo']]
training_p['RegistrationNumber'] = training_p['RegistrationNumber'].fillna(0)
training_p['RegistrationNumber'] = training_p['RegistrationNumber'].astype(int).astype(str).str.zfill(8)
training_p['LastAccessDate'] = pd.to_datetime(training_p['LastAccessDate'])
training_p['CreatedDate'] = pd.to_datetime(training_p['CreatedDate'])
training_p['CompletionDate'] = pd.to_datetime(training_p['CompletionDate'])
training_p.columns = ['cif', 'name', 'base', 'funcao', 'cpf', 'email', 'course_name_p','status_p', 'nota_p', 'ultimo_acesso', 'criacao_acesso', 'data_conclusao_p', 'tipo']
training_p = training_p[['cif', 'course_name_p','status_p', 'nota_p', 'data_conclusao_p']]
training_p = training_p.sort_values(by=['cif','data_conclusao_p'])
training_p = training_p.drop_duplicates(subset = ['cif'], keep = 'last')
training_p

# Dados de Rosters
rosters = funcs.generate_data('data/rosters/')
rosters.cif = rosters.cif.astype(str).str.zfill(8)
rosters

inss = rosters.copy()
inss['dep'] = pd.to_datetime(inss['dep'], format='%Y-%m-%d')
inss['dep'] = inss['dep'].dt.floor('D')
inss = inss[(inss['code'] == 'INSS') & (inss['dep'] == pd.to_datetime(dt.date(dt.now())))]
inss = inss[['cif', 'code']]
inss.columns = ['cif', 'inss']
inss

lnr = rosters.copy()
lnr['dep'] = pd.to_datetime(lnr['dep'], format='%Y-%m-%d')
lnr['dep'] = lnr['dep'].dt.floor('D')
lnr = lnr[(lnr['code'] == 'LIC-SREM') & (lnr['dep'] == pd.to_datetime(dt.date(dt.now())))]
lnr = lnr[['cif', 'code']]
lnr.columns = ['cif', 'lnr']
lnr

rosters_o = rosters[rosters['code'].isin(funcs.training_filter('EMG'))]
rosters_o = rosters_o.sort_values(by=['cif','dep'])
rosters_o = rosters_o[['cif', 'code', 'dep']]
rosters_o.columns = ['cif', 'code_on', 'date_on']
rosters_o = rosters_o.drop_duplicates(subset = ['cif'], keep = 'last')
rosters_o

rosters_p = rosters[rosters['code'].isin(funcs.training_filter('EMG_presencial'))]
rosters_p = rosters_p.sort_values(by=['cif','dep'])
rosters_p = rosters_p[['cif', 'code', 'dep']]
rosters_p.columns = ['cif', 'code_p', 'date_p']
rosters_p = rosters_p.drop_duplicates(subset = ['cif'], keep = 'last')
rosters_p

# Dados de Checks
checks = pd.read_csv('data/checks/check_schedule.csv', sep=';', encoding='utf-8-sig')
checks.cif = checks.cif.astype(str).str.zfill(8)
checks = checks[checks['checkTypeCode'] == training_code]
checks = checks[['cif', 'checkTypeCode', 'expiryDt']]
checks

# Merge dos dados
data_all = pd.merge(crew, checks[['cif', 'checkTypeCode', 'expiryDt']], on='cif', how='left')
data_all = data_all.merge(training, on='cif', how='left')
data_all = data_all.merge(training_p, on='cif', how='left')
data_all = data_all.merge(rosters_o, on='cif', how='left')
data_all = data_all.merge(rosters_p, on='cif', how='left')
data_all = data_all.merge(inss, on='cif', how='left')
data_all = data_all.merge(lnr, on='cif', how='left')
data_all

# Formatação do dataFrame de Dados
validation = data_all.copy()

# Ajuste de data do vencimento
validation['expiryDt'] = pd.to_datetime(validation['expiryDt'], format='%Y-%m-%d')
validation['expiryDt'] = validation['expiryDt'].dt.floor('D')

# Ajuste de data do treinamento online
validation['date_on'] = validation['date_on'].fillna(pd.to_datetime(dt(2035,12,31)))
validation['date_on'] = pd.to_datetime(validation['date_on'], format='%Y-%m-%d')
validation['date_on'] = validation['date_on'].dt.floor('D')

# Ajuste de data do treinamento presencial
validation['date_p'] = validation['date_p'].fillna(pd.to_datetime(dt(2035,12,31)))
validation['date_p'] = pd.to_datetime(validation['date_p'], format='%Y-%m-%d')
validation['date_p'] = validation['date_p'].dt.floor('D')


# Data da conclusao ON LINE
validation['data_conclusao'] = validation['data_conclusao'].fillna(pd.to_datetime(dt(2035,12,31)))
validation['data_conclusao'] = pd.to_datetime(validation['data_conclusao'], format='%Y-%m-%d')
validation['data_conclusao'] = validation['data_conclusao'].dt.floor('D')

# Data da conclusao Presencial
validation['data_conclusao_p'] = validation['data_conclusao_p'].fillna(pd.to_datetime(dt(2035,12,31)))
validation['data_conclusao_p'] = pd.to_datetime(validation['data_conclusao_p'], format='%Y-%m-%d')
validation['data_conclusao_p'] = validation['data_conclusao_p'].dt.floor('D')


validation['presencial_depois_online'] = validation['date_p'] > validation['date_on']
validation = validation[validation['expiryDt'] < (pd.to_datetime(dt.date(dt.now())) + pd.offsets.MonthBegin(0))]
validation = validation[validation['date_p'] < (pd.to_datetime(dt.date(dt.now())))]


validation['tempo_de_conclusao (dias)'] = pd.to_datetime(dt.date(dt.now())) - validation['data_conclusao']

validation['atualizar'] = (validation['status'] == 'Concluído') & \
                          (validation['status_p'] == 'Concluído') & \
                          (validation['presencial_depois_online'])

validation

validation.to_csv(f'./data/output/{training_code}.csv', index=False, sep=';', encoding='utf-8-sig')
